requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
requests==2.32.3
transformers==4.52.4
sentence-transformers==4.1.0
faiss-cpu==1.11.0